just

+++ src/main/java/school/mjc/stage0/loops/task2/PowerOfTwo.java
@@ -7,7 +7,7 @@ public class PowerOfTwo {
         }else {
             int pows = 0;
             while(pows<=power){
-                System.out.println(Math.pow(2,pows));
+                System.out.println((int) Math.pow(2,pows));
                 pows++;
             }
         }
